move "Look & Feel/Main interface" to its own widget
